ProjectC-Polymarketnewrecommendation/requirements.txt
# Polymarket Alpha Scanner - Dependencies

# HTTP & API
httpx==0.27.0
aiohttp==3.9.3

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# AI
openai==1.12.0

# Telegram
python-telegram-bot==21.0.1

# Scheduler
apscheduler==3.10.4

# Web Server
fastapi==0.109.2
uvicorn==0.27.1

# Environment & Config
python-dotenv==1.0.1
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Utilities
rich==13.7.0  # CLI 출력
